Refactor project structure, add 3D interpolation and MIP visualization
